Refactor HO100300
